squadt/{tool_executor,specification}.cpp:  - tools are started properly now  - remaining problem: the same commands are sometimes executed more than once
squadt/libmd5/md5pp.h:
 - fixed typo in union



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@943 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/squadt/libmd5/md5pp.h
@@ -9,8 +9,8 @@
 namespace md5 {
 
   typedef union {
-    uint8_t bytes[16];
-    uint8_t dwords[4];
+    uint8_t  bytes[16];
+    uint32_t dwords[4];
   } compact_digest;
 
   /* Set to the zero checksum */
